+++ src/main/java/kr/ac/bokgpt/api/welfareapi/local/dto/LocalDetailReq.java
@@ -1,4 +1,4 @@
-package kr.ac.bokgpt.welfareapi.local.dto;
+package kr.ac.bokgpt.api.welfareapi.local.dto;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Component;
 @Component
 @Data
 public class LocalDetailReq {
+
     @NotNull
     private String servId;
